Rename: root/types 디렉터리 react-query/hooks/types로 이동
